requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
attrs==25.3.0
audioop-lts==0.2.2
babel==2.17.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
DateTime==5.5
deprecation==2.1.0
discord.py==2.6.4
Flask==3.1.2
frozenlist==1.7.0
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.6.3
outcome==1.3.0.post0
packaging==25.0
postgrest==2.22.2
propcache==0.3.2
py-cord==2.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
realtime==2.22.2
requests==2.32.5
rich==14.1.0
robloxpy==0.2.21
selenium==4.37.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
storage3==2.22.2
StrEnum==0.4.15
supabase==2.22.2
supabase-auth==2.22.2
supabase-functions==2.22.2
tkcalendar==1.6.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
zope.interface==8.0.1
